Integrated rm, shell is mostly working :) need to check for memory errors and heavily test

+++ FS-shell/parser.h
@@ -2,6 +2,6 @@
 #define PARSER_H
 
 char ** tokenize2(char * cmd, int * numCmds);
-u_int8_t * chmodParsing(char * input, int isDir, int * curperms);
+u_int8_t * chmodParsing(char * input, u_int8_t isDir, u_int8_t * curperms);
 char * arrayToPermStr(u_int8_t * perms, int isDir);
 #endif
